refactor: VideoControllerNative & VideoControllerAndroid
