+++ media_kit_video/lib/src/video_controller/native/real.dart
@@ -11,12 +11,12 @@ import 'package:flutter/services.dart';
 import 'package:flutter/foundation.dart';
 import 'package:media_kit/media_kit.dart';
 
-import 'package:media_kit_video/src/video_controller.dart';
+import 'package:media_kit_video/src/video_controller/video_controller.dart';
 
 /// {@template video_controller_native}
 ///
 /// VideoControllerNative
-/// ------------------
+/// ---------------------
 ///
 /// The [VideoController] implementation based on native C/C++ used on:
 /// * Windows
