+++ media_kit_video/lib/src/video_controller/native/stub.dart
@@ -5,33 +5,10 @@
 /// Use of this source code is governed by MIT license that can be found in the LICENSE file.
 
 import 'package:media_kit/media_kit.dart';
-import 'package:media_kit_video/media_kit_video.dart';
 
-// Stub declarations for avoiding dart:ffi import errors on Flutter Web.
+import 'package:media_kit_video/src/video_controller/video_controller.dart';
 
-class VideoControllerAndroid extends VideoController {
-  static const bool supported = false;
-
-  VideoControllerAndroid(
-    super.player,
-    super.width,
-    super.height,
-  );
-
-  static Future<VideoController> create(
-    Player player, {
-    int? width,
-    int? height,
-    bool enableHardwareAcceleration = true,
-  }) =>
-      throw UnimplementedError();
-
-  @override
-  Future<void> dispose() => throw UnimplementedError();
-
-  @override
-  Future<void> setSize({int? width, int? height}) => throw UnimplementedError();
-}
+// Stub declaration for avoiding compilation errors on Dart JS using conditional imports.
 
 class VideoControllerNative extends VideoController {
   static const bool supported = false;
@@ -39,8 +16,9 @@ class VideoControllerNative extends VideoController {
   VideoControllerNative(
     super.player,
     super.width,
-    super.height,
-  );
+    super.height, {
+    super.enableHardwareAcceleration = true,
+  });
 
   static Future<VideoController> create(
     Player player, {
