+++ media_kit_video/lib/src/video_controller/native/video_controller_native.dart
@@ -4,5 +4,4 @@
 /// All rights reserved.
 /// Use of this source code is governed by MIT license that can be found in the LICENSE file.
 
-export 'package:media_kit_video/src/video/video.dart';
-export 'package:media_kit_video/src/video_controller/video_controller.dart';
+export 'real.dart' if (dart.library.html) 'stub.dart';
